Food_Image_Classification/requirements.txt
matplotlib==3.4.3
pillow==8.2.0
transformers==4.17.0
torch==1.10.0